fix: resolve remaining clippy warnings and optimize CI workflow
- Fix derivable_impls in logging/schema.rs and network_transport.rs
- Fix default_constructed_unit_structs in tracing_mod/exporter.rs
- Add safety documentation to unsafe functions in lib.rs
- Remove useless_conversion patterns in ffi_safe.rs
- Remove unnecessary casts in loaders.rs and lifecycle.rs
- Add type alias for complex type in audit.rs
- Fix CI workflow: remove duplicate check job, add job dependencies
- Apply remaining clippy fixes from previous session (and_then, strip_prefix, etc.)

+++ abi/src/lifecycle.rs
@@ -689,7 +689,7 @@ impl PluginLoadPipeline {
     #[allow(dead_code)]
     fn calculate_backoff_delay(&self, retry_count: usize, base_delay_ms: u64) -> u64 {
         // Exponential backoff: 2^retry_count, capped at 2^10 (1024x)
-        let exponential = (base_delay_ms as u64).saturating_mul(1 << (retry_count.min(10)));
+        let exponential = base_delay_ms.saturating_mul(1 << (retry_count.min(10)));
 
         // Add jitter: 0-25% of exponential delay
         // This prevents synchronized retries
